iprop patch (3/4) This patch enables a slave to track changes from the master and commit them to the local ulog.  This facilitates a slave to be promoted as a master and for other slaves to still sync up because their last ulog state can be found.
In essence, when kpropd processes a ulog entry received via iprop, it will
also commit it to the ulog according to the locally configured ulog history
definitions/size to be maintained.

The first DB update when received from a full dump will be set.  Even though
the first entry is not available as an individual record update, it will be
the baseline for future ulog updates.  The first entry will be updated when
the ulog size reaches its limit, after which it will be based on the circular
ulog buffer.