kpropd: Fix KDB ulog mis-tagging when full resync fails.
Sometimes a full resync can fail and leave the ulog with the most current
version despite the failed load (even with the fixes in 1.10.5 & 1.10.6).
If a full resync did not complete successfully, reset the ulog to ensure
the database is not considered current (to ensure future iprop requests
will re-attempt a full resync).

The condition has been seen when the system log contains the entry:
	kpropd: Full resync, invalid return.

The kdb5_util patch in 1.10.5 is insufficient to remediate this condition.