iprop fixes (1/4) - Reset ulog upon policy change
When a policy change is enacted, a ulog reset is required, otherwise, the
slave servers do not receive a new full dump with the new policy definitions.
Principal updates may be referencing a "new" policy which doesn't exist
and fail.

Fix includes a block size edge condition in ulog_resize (if the ulog
entry size fit exactly in the blocksize, there is no need to resize, but
the existing code would have forced a resize)