iprop fixes (2/4) - kproplog
In preparation for another patch, and based on ulog being a circular buffer,
have kproplog loop through the ulog based on the index for the serial number
and assume early entries may be missing. Dereferencing of the ulog should be
consistent with how the KDC dereferences the ulog.

This works with both a ulog where the entry numbers range from m-n where the
number of entries may be smaller than the ulog limit and m need not be 1.
Because the KDC implements dereferencing based on the (sno-1)%size, this is
the correct way.  The current code worked only because the KDC previously
always starts the ulog # at 1, but would fail if the logic were ever changed
to allow another starting index (see next patch).